tcg/backend/requirements.txt
Django==4.2.11
mysqlclient==2.2.4
sqlparse==0.4.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
pillow==10.3.0